IPCAudioDevice now passes generated white noise through audio callback
This was just a test to see if I could observe the buffer data being sent into the EQ plugin and being modifed in the output.

I was able to print out pre and post processing values from the in and out buffers and everything seems kosher. 

Next steps are to set up some form of IPC from the main process to and from the plugin process to transport the audio I/O