IPC audio I/O now working! zeromq + google protobuf implementation is much cleaner than shared memory and seems to be performant enough for now.
The host process now listens for a message containing a protobuf encoded IOBuffer object. Then parses the request and splits up the input and output arrays into multi-dimensional float arrays based on num channels and sample size. This assumes equal # of samples per channel are being provided as input.

Then it passes the copied data through JUCE's audio callback, updates the IOBuffer output data, serializes it and sends it back as a response.