Im moving back to throwing the main api in a native node module. Primary reasons being that the apis for zmq in node do not seem to play nicely in a model that requires synchronous behavior for the req/res responses. Any workarounds I found to getting it done in JS seemed overly convoluted so I'm opting to do all that work in C++ where I won't be fighting the zmq req/res model so much.
So I've removed typescript as a dependency since there won't be any TS or JS in this module anymore, and Re-implemented the api as a no-op interface. running npm-start should playback a wav file if you drop one in the directory named `test.wav` (not tracking it in the repo).

Next steps are to implement the start/stop functions in PluginHost to actually start and stop instance of the host binary.

EXCITING!